Refactoring code:
The install function used to take an 'add' argument to determine whether the filter should be added
to specified profiles or not. The list of profiles was used only when the add flag was set to true.
If user didn't provide the profiles explicitly it was set to a list with a single item - 'default'.
Now, the internal behavior is changed - there is no 'add' flag, but the 'default' profile is
added to the list only if the user uses the --profiles flag in the command. The behavior of the
program from the user point of view is the same. Internally the Install funciton expects
an empty list if no profiles need to be updated.